+++ src/main/java/org/jagar/semaforosmart/repository/DeteccionRTRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DeteccionRTRepository extends JpaRepository<DeteccionRT, Integer> {
-    Optional<DeteccionRT> findTopBySitio_IdOrderByIdDesc(Integer sitioId);
+public interface DeteccionRTRepository extends JpaRepository<DeteccionRT, Long> {
+    Optional<DeteccionRT> findTopBySitio_IdOrderByIdDesc(Long sitioId);
     Optional<DeteccionRT> findTopByOrderByIdDesc();
 }
