+++ src/main/java/org/jagar/semaforosmart/repository/OrdenControlRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrdenControlRepository extends JpaRepository<OrdenControl,Integer> {
+public interface OrdenControlRepository extends JpaRepository<OrdenControl,Long> {
     List<OrdenControl> findBySitio(Sitio sitio);
 }
