+++ src/main/java/org/jagar/semaforosmart/repository/SitioRepository.java
@@ -3,5 +3,5 @@ package org.jagar.semaforosmart.repository;
 import org.jagar.semaforosmart.entity.Sitio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SitioRepository extends JpaRepository<Sitio,Integer> {
+public interface SitioRepository extends JpaRepository<Sitio,Long> {
 }
